Refactor: split ResourceController into focused controllers + ResourceQueryBuilder service
- Extract listing query construction (eager loading, filters, sorting, completeness checks, caching) into App\Services\Resources\ResourceQueryBuilder.
- Move loadMore + getFilterOptions to ResourceFilterController.
- Move exportDataCiteJson + exportDataCiteXml + exportJsonLd to ResourceExportController.
- Move registerDoi + getDataCitePrefixes to ResourceDoiRegistrationController.
- Slim ResourceController down to index/store/storeDraft/destroy/destroyAll.
- Update routes/web.php controller references (URLs and route names unchanged).
- Regenerate Wayfinder.
- Whitelist ResourceQueryBuilder in the Service-suffix arch test.
